+++ back-end/src/main/java/com/sept/backend/controller/UserController.java
@@ -1,7 +1,7 @@
 package com.sept.backend.controller;
 
 import com.sept.backend.exception.Status;
-import com.sept.backend.model.AuthRequest;
+import com.sept.backend.payload.UserLoginRequest;
 import com.sept.backend.model.User;
 import com.sept.backend.model.Users;
 import com.sept.backend.payload.UserUpdateStatusRequest;
@@ -20,12 +20,8 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.client.RestTemplate;
 
-import javax.print.Doc;
 import javax.validation.Valid;
-import java.io.OutputStream;
-import java.net.HttpURLConnection;
 import java.net.URL;
-import java.net.URLConnection;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -62,12 +58,12 @@ public class UserController {
     }
 
     @PostMapping("/users/login")
-    public ResponseEntity<?> loginUserWithJWT(@RequestBody AuthRequest authRequest) throws Exception {
+    public ResponseEntity<?> loginUserWithJWT(@RequestBody UserLoginRequest userLoginRequest) throws Exception {
         // authenticate user
-        authenticate(authRequest.getEmail(), authRequest.getPassword());
-        final UserDetails userDetails = userDetailService.loadUserByUsername(authRequest.getEmail());
+        authenticate(userLoginRequest.getEmail(), userLoginRequest.getPassword());
+        final UserDetails userDetails = userDetailService.loadUserByUsername(userLoginRequest.getEmail());
         // generate token
-        final String token = jwtUtil.generateToken(authRequest.getEmail());
+        final String token = jwtUtil.generateToken(userLoginRequest.getEmail());
         // return jwt token
         Map<String,Object> jwtResponse = new HashMap<>();
         jwtResponse.put("token", token);
